refactor: hook내 queries 분류
