+++ breaking-front/src/pages/Profile/Profile.js
@@ -2,10 +2,10 @@ import Line from 'components/Line/Line';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
 import { ProfileSkeleton } from 'components/Skeleton/Skeleton';
 import Tabs from 'components/Tabs/Tabs';
-import useProfile from 'pages/Profile/hooks/useProfile';
-import useProfileBookmarkedPost from 'pages/Profile/hooks/useProfileBookmarkedPost';
-import useProfileBoughtPost from 'pages/Profile/hooks/useProfileBoughtPost';
-import useProfileWrittenPost from 'pages/Profile/hooks/useProfileWrittenPost';
+import useProfile from 'pages/Profile/hooks/queries/useProfile';
+import useProfileBookmarkedPost from 'pages/Profile/hooks/queries/useProfileBookmarkedPost';
+import useProfileBoughtPost from 'pages/Profile/hooks/queries/useProfileBoughtPost';
+import useProfileWrittenPost from 'pages/Profile/hooks/queries/useProfileWrittenPost';
 import useCheckMyPage from 'hooks/useCheckMyPage';
 import * as Style from 'pages/Profile/Profile.styles';
 import React, { useState } from 'react';
